Move RHContributionEditableBase to base.py
